+++ iroh-services-proto/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "iroh-services-proto"
+version = "0.1.0"
+edition = "2024"
+description = "Shared protocol types for iroh-services"
+license = "MIT OR Apache-2.0"
+authors = ["n0 team"]
+repository = "https://github.com/n0-computer/iroh-services"
+rust-version = "1.91"
+
+[dependencies]
+anyhow = "1.0.95"
+derive_more = { version = "2.0.1", features = ["display", "from"] }
+irpc = { version = "0.17", default-features = false, features = ["derive", "rpc", "stream", "spans"] }
+iroh = { version = "1.0.0", default-features = false, features = ["unstable-net-report"] }
+iroh-metrics = { version = "1.0.0", default-features = false }
+rcan = "0.4"
+serde = { version = "1.0.217", features = ["derive"] }
+strum = { version = "0.28", features = ["derive"] }
+thiserror = "2.0.12"
+tracing = "0.1.41"
+uuid = { version = "1.17", features = ["serde"] }
+
+[dev-dependencies]
+ed25519-dalek = "=3.0.0-rc.0"
+postcard = { version = "1.1.3", features = ["use-std"] }
