+++ iroh-services-proto/src/lib.rs
@@ -1,0 +1,5 @@
+//! Shared capability and wire-protocol types for iroh-services clients and servers.
+
+pub mod caps;
+pub mod net_diagnostics;
+pub mod protocol;
