+++ iroh-services-proto/src/protocol.rs
@@ -1,0 +1,242 @@
+use std::collections::BTreeMap;
+
+use anyhow::Result;
+use irpc::{channel::oneshot, rpc_requests};
+use rcan::Rcan;
+use serde::{Deserialize, Serialize};
+use uuid::Uuid;
+
+use crate::{caps::Caps, net_diagnostics::DiagnosticsReport};
+
+/// The main ALPN for connecting from the client to the cloud node.
+///
+/// # Versioning
+///
+/// The wire protocol is append-only and does not bump this ALPN for additive
+/// changes. postcard encodes enum variants by index, so as long as new
+/// [`IrohServicesProtocol`] and [`RemoteError`] variants are only appended
+/// (never inserted, reordered, or removed), older messages stay wire-compatible:
+///
+/// - An older client always works against a newer server: the server decodes
+///   every request the client can send, and only replies with error variants the
+///   client already knows.
+/// - A newer client against an older server keeps working for the pre-existing
+///   requests (auth, metrics, and so on); a request the old server does not know
+///   fails as a per-request error rather than breaking the connection.
+///
+/// The cloud node is deployed at or ahead of the clients that talk to it, so the
+/// second case is transient and limited to the new calls. A breaking change
+/// (reordering or removing variants, or changing a message's shape) requires a
+/// new ALPN.
+pub const ALPN: &[u8] = b"/iroh/n0des/1";
+
+/// The ALPN for sending messages from the cloud node to the client.
+pub const CLIENT_HOST_ALPN: &[u8] = b"n0/n0des-client-host/1";
+
+pub type IrohServicesClient = irpc::Client<IrohServicesProtocol>;
+
+pub type ClientHostClient = irpc::Client<ClientHostProtocol>;
+
+/// New request variants MUST be appended, never inserted or reordered. See the
+/// versioning policy on [`ALPN`].
+#[rpc_requests(message = ServicesMessage)]
+#[derive(Debug, Serialize, Deserialize)]
+#[allow(clippy::large_enum_variant)]
+#[non_exhaustive]
+pub enum IrohServicesProtocol {
+    #[rpc(tx=oneshot::Sender<()>)]
+    Auth(Auth),
+    #[rpc(tx=oneshot::Sender<RemoteResult<()>>)]
+    PutMetrics(PutMetrics),
+    #[rpc(tx=oneshot::Sender<Pong>)]
+    Ping(Ping),
+
+    #[rpc(tx=oneshot::Sender<RemoteResult<()>>)]
+    PutNetworkDiagnostics(PutNetworkDiagnostics),
+
+    #[rpc(tx=oneshot::Sender<RemoteResult<()>>)]
+    GrantCap(GrantCap),
+
+    #[rpc(tx=oneshot::Sender<RemoteResult<()>>)]
+    NameEndpoint(NameEndpoint),
+
+    #[rpc(tx=oneshot::Sender<RemoteResult<()>>)]
+    SetGroup(SetGroup),
+
+    #[rpc(tx=oneshot::Sender<RemoteResult<()>>)]
+    SetAttributes(SetAttributes),
+}
+
+/// Dedicated protocol for cloud-to-endpoint net diagnostics connections.
+#[rpc_requests(message = NetDiagnosticsMessage)]
+#[derive(Debug, Serialize, Deserialize)]
+#[allow(clippy::large_enum_variant)]
+pub enum ClientHostProtocol {
+    #[rpc(tx=oneshot::Sender<()>)]
+    Auth(Auth),
+    #[rpc(tx=oneshot::Sender<RemoteResult<DiagnosticsReport>>)]
+    RunNetworkDiagnostics(RunNetworkDiagnostics),
+}
+
+pub type RemoteResult<T> = Result<T, RemoteError>;
+
+#[derive(Clone, Serialize, Deserialize, thiserror::Error, Debug)]
+#[non_exhaustive]
+pub enum RemoteError {
+    // The first three variants and their order are the v1 wire contract: postcard
+    // encodes enum variants by index, so a v1 client only decodes these and at
+    // their original positions. New variants MUST be appended after them, and the
+    // server must only send new variants in response to new (v2+) requests.
+    #[error("Missing capability: {}", _0.to_strings().join(", "))]
+    MissingCapability(Caps),
+    #[error("Unauthorized: {}", _0)]
+    AuthError(String),
+    #[error("Internal server error")]
+    InternalServerError,
+    #[error("Invalid input: {}", _0)]
+    InvalidInput(String),
+    #[error("Rate limit exceeded")]
+    RateLimited,
+}
+
+/// Authentication on first request.
+#[derive(Debug, derive_more::Display, Serialize, Deserialize)]
+#[display("Auth")]
+pub struct Auth {
+    pub caps: Rcan<Caps>,
+}
+
+/// Request to store the given metrics data.
+#[derive(Debug, derive_more::Display, Serialize, Deserialize)]
+#[display("PutMetrics")]
+pub struct PutMetrics {
+    pub session_id: Uuid,
+    pub update: iroh_metrics::encoding::Update,
+}
+
+/// Simple ping request.
+#[derive(Debug, derive_more::Display, Serialize, Deserialize)]
+#[display("Ping")]
+pub struct Ping {
+    pub req_id: [u8; 16],
+}
+
+/// Simple ping response.
+#[derive(Debug, Serialize, Deserialize)]
+pub struct Pong {
+    pub req_id: [u8; 16],
+}
+
+/// Publishes network diagnostics.
+#[derive(Debug, derive_more::Display, Serialize, Deserialize)]
+#[display("PutNetworkDiagnostics")]
+pub struct PutNetworkDiagnostics {
+    pub report: DiagnosticsReport,
+}
+
+/// Asks this node to run diagnostics and return the result.
+#[derive(Debug, derive_more::Display, Serialize, Deserialize)]
+#[display("RunNetworkDiagnostics")]
+pub struct RunNetworkDiagnostics;
+
+/// Grants a capability token to the remote endpoint.
+#[derive(Debug, derive_more::Display, Serialize, Deserialize)]
+#[display("GrantCap")]
+pub struct GrantCap {
+    pub cap: Rcan<Caps>,
+}
+
+/// Labels the client endpoint cloud-side with a string identifier.
+#[derive(Debug, derive_more::Display, Serialize, Deserialize)]
+#[display("NameEndpoint")]
+pub struct NameEndpoint {
+    pub name: String,
+}
+
+/// Attach the client endpoint to a single named group cloud-side.
+#[derive(Debug, derive_more::Display, Serialize, Deserialize)]
+#[display("SetGroup")]
+pub struct SetGroup {
+    pub group: String,
+}
+
+/// Replace the arbitrary key-value attributes on the client endpoint cloud-side.
+#[derive(Debug, derive_more::Display, Serialize, Deserialize)]
+#[display("SetAttributes")]
+pub struct SetAttributes {
+    pub attributes: BTreeMap<String, String>,
+}
+
+#[cfg(test)]
+mod tests {
+    use std::collections::BTreeMap;
+
+    use ed25519_dalek::SigningKey;
+    use rcan::Expires;
+
+    use super::*;
+    use crate::caps::RelayCap;
+
+    #[test]
+    fn auth_wire_encoding_is_stable() {
+        const GOLDEN: &[u8] = &[
+            32, 138, 136, 227, 221, 116, 9, 241, 149, 253, 82, 219, 45, 60, 186, 93, 114, 202, 103,
+            9, 191, 29, 148, 18, 27, 243, 116, 136, 1, 180, 15, 111, 92, 32, 129, 57, 119, 14, 168,
+            125, 23, 95, 86, 163, 84, 102, 195, 76, 126, 204, 203, 141, 138, 145, 180, 238, 55,
+            162, 93, 246, 15, 91, 143, 201, 179, 148, 0, 0, 1, 2, 0, 1, 128, 164, 167, 218, 6, 177,
+            237, 32, 146, 189, 68, 211, 5, 10, 199, 42, 240, 27, 8, 241, 35, 217, 85, 253, 169,
+            122, 177, 208, 76, 202, 130, 200, 245, 37, 86, 179, 157, 6, 0, 247, 89, 83, 57, 23,
+            236, 89, 175, 142, 180, 69, 198, 251, 51, 212, 249, 182, 228, 156, 226, 10, 172, 167,
+            139, 111, 161, 120, 190, 235, 10,
+        ];
+
+        let issuer = SigningKey::from_bytes(&[1; 32]);
+        let audience = SigningKey::from_bytes(&[2; 32]).verifying_key();
+        let caps = Rcan::issuing_builder(&issuer, audience, Caps::new([RelayCap::Use]))
+            .sign(Expires::At(1_800_000_000));
+        let auth = Auth { caps };
+
+        assert_eq!(postcard::to_stdvec(&auth).unwrap(), GOLDEN);
+        postcard::from_bytes::<Auth>(GOLDEN).unwrap();
+    }
+
+    #[test]
+    fn remote_error_wire_encoding_is_stable() {
+        let idx = |error: &RemoteError| postcard::to_stdvec(error).expect("encode")[0];
+        assert_eq!(idx(&RemoteError::AuthError(String::new())), 1);
+        assert_eq!(idx(&RemoteError::InternalServerError), 2);
+        assert_eq!(idx(&RemoteError::InvalidInput(String::new())), 3);
+        assert_eq!(idx(&RemoteError::RateLimited), 4);
+    }
+
+    #[test]
+    fn set_group_roundtrips() {
+        for group in ["staging", "my-group 👋"] {
+            let message = SetGroup {
+                group: group.to_string(),
+            };
+            let bytes = postcard::to_stdvec(&message).expect("serialize");
+            let decoded: SetGroup = postcard::from_bytes(&bytes).expect("deserialize");
+            assert_eq!(decoded.group, message.group);
+        }
+    }
+
+    #[test]
+    fn set_attributes_roundtrips() {
+        let empty = SetAttributes {
+            attributes: BTreeMap::new(),
+        };
+        let bytes = postcard::to_stdvec(&empty).expect("serialize");
+        let decoded: SetAttributes = postcard::from_bytes(&bytes).expect("deserialize");
+        assert_eq!(decoded.attributes, empty.attributes);
+
+        let mut attributes = BTreeMap::new();
+        attributes.insert("région 🌍".to_string(), "us-wëst 🚀".to_string());
+        attributes.insert("max".to_string(), "x".repeat(128));
+
+        let message = SetAttributes { attributes };
+        let bytes = postcard::to_stdvec(&message).expect("serialize");
+        let decoded: SetAttributes = postcard::from_bytes(&bytes).expect("deserialize");
+        assert_eq!(decoded.attributes, message.attributes);
+    }
+}
